app/requirements.txt
PyYAML==3.12
pytest==3.4.2